+++ datumaro/components/dataset.py
@@ -14,6 +14,7 @@ import os
 import os.path as osp
 import shutil
 
+from datumaro.components.annotation import AnnotationType, LabelCategories
 from datumaro.components.dataset_filter import (
     XPathAnnotationsFilter, XPathDatasetFilter,
 )
@@ -23,8 +24,8 @@ from datumaro.components.errors import (
     RepeatedItemError, UnknownFormatError,
 )
 from datumaro.components.extractor import (
-    DEFAULT_SUBSET_NAME, AnnotationType, CategoriesInfo, DatasetItem, Extractor,
-    IExtractor, ItemTransform, LabelCategories, Transform,
+    DEFAULT_SUBSET_NAME, CategoriesInfo, DatasetItem, Extractor, IExtractor,
+    ItemTransform, Transform,
 )
 from datumaro.util import error_rollback, is_method_redefined, on_error_do
 from datumaro.util.log_utils import logging_disabled