+++ datumaro/plugins/cvat_format/converter.py
@@ -9,11 +9,10 @@ import logging as log
 import os
 import os.path as osp
 
+from datumaro.components.annotation import AnnotationType, LabelCategories
 from datumaro.components.converter import Converter
 from datumaro.components.dataset import ItemStatus
-from datumaro.components.extractor import (
-    AnnotationType, DatasetItem, LabelCategories,
-)
+from datumaro.components.extractor import DatasetItem
 from datumaro.util import cast, pairs
 
 from .format import CvatPath