+++ datumaro/plugins/icdar_format/converter.py
@@ -5,8 +5,8 @@
 import os
 import os.path as osp
 
+from datumaro.components.annotation import AnnotationType, CompiledMask
 from datumaro.components.converter import Converter
-from datumaro.components.extractor import AnnotationType, CompiledMask
 from datumaro.util.image import save_image
 from datumaro.util.mask_tools import paint_mask