+++ datumaro/plugins/open_images_format.py
@@ -19,15 +19,15 @@ from attr import attrs
 import cv2
 import numpy as np
 
+from datumaro.components.annotation import (
+    AnnotationType, Bbox, Label, LabelCategories, Mask,
+)
 from datumaro.components.converter import Converter
 from datumaro.components.dataset import ItemStatus
 from datumaro.components.errors import (
     DatasetError, RepeatedItemError, UndefinedLabel,
 )
-from datumaro.components.extractor import (
-    AnnotationType, Bbox, DatasetItem, Extractor, Importer, Label,
-    LabelCategories, Mask,
-)
+from datumaro.components.extractor import DatasetItem, Extractor, Importer
 from datumaro.components.validator import Severity
 from datumaro.util.annotation_util import find_instances
 from datumaro.util.image import (
@@ -484,7 +484,7 @@ class OpenImagesExtractor(Extractor):
         raw = load_image(path, dtype=np.uint8)
         resized = cv2.resize(raw, (size[1], size[0]),
             interpolation=cv2.INTER_NEAREST)
-        return resized.astype(np.bool)
+        return resized.astype(bool)
 
 class OpenImagesImporter(Importer):
     @classmethod