+++ tests/test_imagenet_txt_format.py
@@ -3,10 +3,11 @@ import os.path as osp
 
 import numpy as np
 
-from datumaro.components.dataset import Dataset
-from datumaro.components.extractor import (
-    AnnotationType, DatasetItem, Label, LabelCategories,
+from datumaro.components.annotation import (
+    AnnotationType, Label, LabelCategories,
 )
+from datumaro.components.dataset import Dataset
+from datumaro.components.extractor import DatasetItem
 from datumaro.plugins.imagenet_txt_format import (
     ImagenetTxtConverter, ImagenetTxtImporter,
 )
@@ -119,7 +120,10 @@ class ImagenetTxtFormatTest(TestCase):
             compare_datasets(self, dataset, parsed_dataset,
                 require_images=True)
 
-DUMMY_DATASET_DIR = osp.join(osp.dirname(__file__), 'assets', 'imagenet_txt_dataset')
+DUMMY_DATASET_DIR = osp.join(
+    osp.dirname(__file__), 'assets/imagenet_txt_dataset/basic')
+DUMMY_DATASET_WITH_CUSTOM_LABELS_DIR = osp.join(
+    osp.dirname(__file__), 'assets/imagenet_txt_dataset/custom_labels')
 
 class ImagenetTxtImporterTest(TestCase):
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
@@ -143,6 +147,20 @@ class ImagenetTxtImporterTest(TestCase):
         })
 
         dataset = Dataset.import_from(DUMMY_DATASET_DIR, 'imagenet_txt')
+
+        compare_datasets(self, expected_dataset, dataset, require_images=True)
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_can_import_with_custom_labels_file(self):
+        expected_dataset = Dataset.from_iterable([
+            DatasetItem(id='1', subset='train',
+                annotations=[Label(0)]
+            ),
+        ], categories=['alt_label_%s' % label for label in range(10)])
+
+        dataset = Dataset.import_from(
+            DUMMY_DATASET_WITH_CUSTOM_LABELS_DIR, 'imagenet_txt',
+            labels='synsets-alt.txt')
 
         compare_datasets(self, expected_dataset, dataset, require_images=True)