+++ tests/test_labelme_format.py
@@ -5,8 +5,9 @@ import os.path as osp
 
 import numpy as np
 
+from datumaro.components.annotation import Bbox, Mask, Polygon
 from datumaro.components.dataset import Dataset
-from datumaro.components.extractor import Bbox, DatasetItem, Mask, Polygon
+from datumaro.components.extractor import DatasetItem
 from datumaro.plugins.labelme_format import LabelMeConverter, LabelMeImporter
 from datumaro.util.image import Image
 from datumaro.util.test_utils import (