+++ tests/test_lfw_format.py
@@ -3,8 +3,9 @@ import os.path as osp
 
 import numpy as np
 
+from datumaro.components.annotation import Label, Points
 from datumaro.components.dataset import Dataset
-from datumaro.components.extractor import DatasetItem, Label, Points
+from datumaro.components.extractor import DatasetItem
 from datumaro.plugins.lfw_format import LfwConverter, LfwImporter
 from datumaro.util.image import Image
 from datumaro.util.test_utils import TestDir, compare_datasets